requirements.txt
# ── Core async runtime ────────────────────────────────────────────────────────
asyncio-throttle==1.0.2
uvloop==0.19.0

# ── Exchange connectivity ─────────────────────────────────────────────────────
ccxt==4.2.25
websockets==12.0
aiohttp==3.9.3

# ── Data & numerics ───────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
scikit-learn==1.4.1.post1
statsmodels==0.14.1

# ── Technical analysis ────────────────────────────────────────────────────────
ta-lib==0.4.28
pandas-ta==0.3.14b0

# ── Machine learning ──────────────────────────────────────────────────────────
lightgbm==4.3.0
xgboost==2.0.3
torch==2.2.1
joblib==1.3.2

# ── Web / API ─────────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.28.0
httpx==0.27.0
pydantic==2.6.3
python-telegram-bot==21.1.1

# ── Storage ───────────────────────────────────────────────────────────────────
asyncpg==0.29.0
redis[hiredis]==5.0.3
sqlalchemy==2.0.29
alembic==1.13.1

# ── Observability ─────────────────────────────────────────────────────────────
prometheus-client==0.20.0
loguru==0.7.2
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0

# ── Configuration ─────────────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1
dynaconf==3.2.4

# ── Serialization / inter-service ────────────────────────────────────────────
protobuf==4.25.3
grpcio==1.62.1
grpcio-tools==1.62.1

# ── Rust FFI build ────────────────────────────────────────────────────────────
maturin==1.4.0

# ── Jupyter / research ───────────────────────────────────────────────────────
jupyterlab==4.1.3
ipywidgets==8.1.2
nbstripout==0.7.1
plotly==5.20.0
kaleido==0.2.1

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity==8.2.3
orjson==3.9.15
msgpack==1.0.8
python-dateutil==2.9.0.post0
pytz==2024.1
tqdm==4.66.2
tabulate==0.9.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
pytest-cov==5.0.0
hypothesis==6.100.1
